clicmd: Hide restricted functions by role
Introduce a new keyword `@restrict` in comment blocks
preceding the function definitions.

Check the caller's role (membership in `priv-*` group),
and only import the functions allowed for that group
according to the value of `@restrict`.

That allows for autocomplete to work properly and prevents
sudo password request from appearing when a restricted
subcommand is invoked by unprivileged user.

Also check the role in print_help and only display the
commands and sub-commands allowed for the caller's role.

Signed-off-by: Alexander Amelkin <a.amelkin@yadro.com>