install: Prohibit sudo by default, no lecture
If a user ran a CLI command that requires `sudo`, but they
didn't have a permission (although this should never happen),
or the user directly used `sudo`, they would normally get
a standard lecture on the rules of `sudo` usage, and then
asked for their password.

With YADRO CLI installed, users aren't allowed to run anything
with `sudo` directly. There is no point in lecturing them.

This commit disables lecturing and also disables asking for
a password, because there is no point in asking when running
the command would anyway be denied.

Signed-off-by: Alexander Amelkin <a.amelkin@yadro.com>