bmc: Fix help for ifconfig and datetime ntpconfig
Requires latest version of obmc-yadro-netconfig.
For vegman, remove the local help from the `bmc datetime ntpconfig`
function description block and implement calling `netconfig` for help.
For nicole, update the local help to be up to date with `netconfig`.

For all targets pass the CLI command name as argv[0] to
`netconfig` and call it with the new `--cli` option so that
`netconfig` displays the CLI command instead of its own name.

End-user-impact: CLI: Help for `bmc ifconfig` and
                 `bmc datetime ntpconfig` has been fixed.

Signed-off-by: Alexander Amelkin <a.amelkin@yadro.com>