help: Restrict top-level commands
Allow for some top-level commands to be hidden from
some system roles. If there is a '# Restrict: <roles>'
line in the command script, the command will only be
available to those roles. Other roles will not see it
in the top level `help` and will get a `command not found`
when trying to run it.

End-user-impact: None
Signed-off-by: Alexander Amelkin <a.amelkin@yadro.com>