Fix error handling in local variable assignments
When a variable is assigned a value printed by a function
that can fail and return a non-zero value, and a `set -e`
is in effect, it is expected that such an assignment will
fail. However since `local` is a command per se and it
doesn't fail, the whole expression is considered successful.

Thus, in order to handle errors properly in such assignments,
I here separate declaration of a `local` variable from its
definition.

Signed-off-by: Alexander Amelkin <a.amelkin@yadro.com>