bmc datetime set: Fix format checking
Shellcheck complained about some wrong escaping in
what was thought to be the regexp. While checking that,
it turned out that it wasn't considered as a regexp
by bash at all, and also a negation was missing in
the expression.

That was the source of the known "`bmc datetime set` can
accept any value" problem and is fixed now.

Signed-off-by: Alexander Amelkin <a.amelkin@yadro.com>