clicmd: Add a generic exec_tool function
Now, as hostpwrctl is added as another external tool
that may display its own help, remove `call_netconfig`
and add a generic `exec_tool` function instead. Use it
both for `netconfig` and for `hostpwrctl` to substitute
their names by the name of the called CLI command.

Also use it for `fwupdate` for uniformity, although that
tool is called in CLI in a way that should never result
in printing of usage help by the tool itself.

Signed-off-by: Alexander Amelkin <a.amelkin@yadro.com>