bmc datetime set: Add range sanity checks
Only allow valid dates. The year range is limited 2021..2037.
The lower limit is obvious. The upper limit is because
mktime() in OpenBMC suffers from Y2k38 problem and won't accept
dates beyond 03:14:07 on Tuesday, 19 January 2038, so 2037
is the last year for which the full range of other components
is accepted.

End-user-impact: Only valid dates are now allowed
Signed-off-by: Alexander Amelkin <a.amelkin@yadro.com>