host:virtualmedia: Move busctl to vm-cli
Current host virtualmedia implementation based on the `busctl` that
doesn't have a valid mechanism to transfer the `Unix-fd` of UNIX pipes.
That is following to impossible to specify mounting credentials.

That patch brings a way to process virtual-media operations via the
special `vm-cli` tool.

End-user-impact: Now the `host virtualmedia mount` command accepts
                 credentials properly.

Signed-off-by: Igor Kononenko <i.kononenko@yadro.com>